+++ gluten-core/src/main/scala/org/apache/spark/sql/execution/datasources/GlutenColumnarRules.scala
@@ -35,53 +35,6 @@ import org.apache.spark.sql.types.{DataType, Decimal}
 import org.apache.spark.sql.vectorized.ColumnarBatch
 import org.apache.spark.unsafe.types.{CalendarInterval, UTF8String}
 
-class FakeRow(val batch: ColumnarBatch) extends InternalRow {
-  override def numFields: Int = throw new UnsupportedOperationException()
-
-  override def setNullAt(i: Int): Unit = throw new UnsupportedOperationException()
-
-  override def update(i: Int, value: Any): Unit = throw new UnsupportedOperationException()
-
-  override def copy(): InternalRow = throw new UnsupportedOperationException()
-
-  override def isNullAt(ordinal: Int): Boolean = throw new UnsupportedOperationException()
-
-  override def getBoolean(ordinal: Int): Boolean = throw new UnsupportedOperationException()
-
-  override def getByte(ordinal: Int): Byte = throw new UnsupportedOperationException()
-
-  override def getShort(ordinal: Int): Short = throw new UnsupportedOperationException()
-
-  override def getInt(ordinal: Int): Int = throw new UnsupportedOperationException()
-
-  override def getLong(ordinal: Int): Long = throw new UnsupportedOperationException()
-
-  override def getFloat(ordinal: Int): Float = throw new UnsupportedOperationException()
-
-  override def getDouble(ordinal: Int): Double = throw new UnsupportedOperationException()
-
-  override def getDecimal(ordinal: Int, precision: Int, scale: Int): Decimal =
-    throw new UnsupportedOperationException()
-
-  override def getUTF8String(ordinal: Int): UTF8String =
-    throw new UnsupportedOperationException()
-
-  override def getBinary(ordinal: Int): Array[Byte] = throw new UnsupportedOperationException()
-
-  override def getInterval(ordinal: Int): CalendarInterval =
-    throw new UnsupportedOperationException()
-
-  override def getStruct(ordinal: Int, numFields: Int): InternalRow =
-    throw new UnsupportedOperationException()
-
-  override def getArray(ordinal: Int): ArrayData = throw new UnsupportedOperationException()
-
-  override def getMap(ordinal: Int): MapData = throw new UnsupportedOperationException()
-
-  override def get(ordinal: Int, dataType: DataType): AnyRef =
-    throw new UnsupportedOperationException()
-}
-
 case class ColumnarToFakeRowStrategy(session: SparkSession) extends Strategy {
   override def apply(plan: LogicalPlan): Seq[SparkPlan] =
     LogicalPlanSelector.maybeNil(session, plan) {
