+++ package/pom.xml
@@ -46,11 +46,6 @@
           <artifactId>backends-velox</artifactId>
           <version>${project.version}</version>
         </dependency>
-        <dependency>
-          <groupId>io.glutenproject</groupId>
-          <artifactId>spark-sql-columnar-shims-write</artifactId>
-          <version>${project.version}</version>
-        </dependency>
       </dependencies>
     </profile>
     <profile>
