+++ shims/pom.xml
@@ -57,7 +57,6 @@
         <activeByDefault>true</activeByDefault>
       </activation>
       <modules>
-        <module>write</module>
         <module>common</module>
         <module>spark32</module>
       </modules>
@@ -65,7 +64,6 @@
     <profile>
       <id>spark-3.3</id>
       <modules>
-        <module>write</module>
         <module>common</module>
         <module>spark33</module>
       </modules>
